feat(client): session token 64-bit + auth UI tweaks (match server)
- ClientSession.Token + MasterClient.Token -> ulong (handshake / login /
  server-select carry the widened 64-bit token; ProtocolVersion v2).
- AuthScreen: generic "user or email already registered" message
  (anti-enumeration) + password >=8 hint.
- Refresh Plugins DLLs (ArcaneShared + NetworkLibrary) to the 64-bit /
  protocol-v2 build (both must move together — NetworkLibrary defines ReadULong).

+++ Assets/Arcane_Aegis/Scripts/Network/ClientSession.cs
@@ -8,7 +8,7 @@ namespace Arcane_Aegis.Network
     public static class ClientSession
     {
         public static uint AccountId;
-        public static uint Token;
+        public static ulong Token;
         public static byte ServerId;
         public static uint CharacterId; // chosen character (set on select → used to enter the world)
     }
